+++ v2/pkg/engine/datasource/grpc_datasource/compiler.go
@@ -399,6 +399,8 @@ type ServiceCall struct {
 	Output protoref.Message
 	// RPC is the call that was made to the gRPC service
 	RPC *RPCCall
+	// Skip is true if the service call should not be invoked
+	Skip bool
 }
 
 func (s *ServiceCall) MethodFullName() string {
@@ -464,6 +466,10 @@ func (p *RPCCompiler) CompileFetches(graph *DependencyGraph, fetches []FetchItem
 			return nil, err
 		}
 
+		if serviceCall.Skip {
+			continue
+		}
+
 		graph.SetFetchData(node.ID, &serviceCall)
 		serviceCalls = append(serviceCalls, serviceCall)
 	}
@@ -508,8 +514,10 @@ func (p *RPCCompiler) CompileNode(graph *DependencyGraph, fetch FetchItem, input
 		}
 
 		request, err = p.buildProtoMessageWithContext(inputMessage, &call.Request, inputData, context)
-		if err != nil {
-			return ServiceCall{}, err
+		if err != nil || request == nil {
+			return ServiceCall{
+				Skip: true,
+			}, err
 		}
 	}
 
@@ -572,6 +580,12 @@ func (p *RPCCompiler) buildProtoMessageWithContext(inputMessage Message, rpcMess
 		return nil, fmt.Errorf("context is required for resolve calls")
 	}
 
+	// If the context is empty, we can skip the invocation
+	// Currently only one context is supported for resolve calls. This might be extended in the future.
+	if context[0].ServiceCall == nil || context[0].ServiceCall.Output == nil {
+		return nil, nil
+	}
+
 	if p.doc.MessageRefByName(rpcMessage.Name) == InvalidRef {
 		return nil, fmt.Errorf("message %s not found in document", rpcMessage.Name)
 	}
@@ -597,9 +611,13 @@ func (p *RPCCompiler) buildProtoMessageWithContext(inputMessage Message, rpcMess
 		return nil, fmt.Errorf("context field not found in message %s", inputMessage.Name)
 	}
 
-	contextList := p.newEmptyListMessageByName(rootMessage, contextSchemaField.Name)
-	contextData := p.resolveContextData(context[0], contextRPCField) // TODO handle multiple contexts (resolver requires another resolver)
+	// TODO handle multiple contexts (resolver requires another resolver)
+	contextData := p.resolveContextData(context[0], contextRPCField)
+	if len(contextData) == 0 {
+		return nil, nil
+	}
 
+	contextList := p.newEmptyListMessageByName(rootMessage, contextSchemaField.Name)
 	for _, contextElement := range contextData {
 		val := contextList.NewElement()
 		valMsg := val.Message()
@@ -680,6 +698,10 @@ func (p *RPCCompiler) resolveContextDataForPath(message protoref.Message, path a
 
 // resolveListDataForPath resolves the data for a given path in a list message.
 func (p *RPCCompiler) resolveListDataForPath(message protoref.List, fd protoref.FieldDescriptor, path ast.Path) []protoref.Value {
+	if !message.IsValid() {
+		return nil
+	}
+
 	if path.Len() == 0 {
 		return nil
 	}
@@ -713,6 +735,10 @@ func (p *RPCCompiler) resolveListDataForPath(message protoref.List, fd protoref.
 
 // resolveDataForPath resolves the data for a given path in a message.
 func (p *RPCCompiler) resolveDataForPath(messsage protoref.Message, path ast.Path) []protoref.Value {
+	if !messsage.IsValid() {
+		return nil
+	}
+
 	if path.Len() == 0 {
 		return nil
 	}
@@ -738,7 +764,15 @@ func (p *RPCCompiler) resolveDataForPath(messsage protoref.Message, path ast.Pat
 	switch fd.Kind() {
 	case protoref.MessageKind:
 		if fd.IsList() {
+			if !field.List().IsValid() {
+				return nil
+			}
+
 			return []protoref.Value{protoref.ValueOfList(field.List())}
+		}
+
+		if !field.Message().IsValid() {
+			return nil
 		}
 
 		return p.resolveDataForPath(field.Message(), path[1:])